requirements.txt
pandas>=2.0.0
requests>=2.31.0
streamlit>=1.32.0
python-dotenv>=1.0.0
plotly>=5.20.0
sqlalchemy>=2.0.0
pytest>=8.0.0
